fix : directory 분류 및 코드리뷰 이후  리팩토링
