+++ src/Component/Footer/Link.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Link = ({ href, text }) => (
+  <a className="footer-link" href={href}>
+    {text}
+  </a>
+);
+
+export default Link;
